+++ lib/pokemon_gen2impl.hpp
@@ -17,22 +17,22 @@ namespace pkmn
     {
         public:
 
-            pokemon_gen2impl(uint16_t species, uint16_t version,
-                             uint8_t level,
-                             uint8_t move1, uint8_t move2,
-                             uint8_t move3, uint8_t move4);
+            pokemon_gen2impl(int species, int version,
+                             int level,
+                             int move1, int move2,
+                             int move3, int move4);
             pokemon_gen2impl(const pkmn::native::gen2_pc_pokemon_t& raw,
-                             uint8_t version);
+                             int version);
             pokemon_gen2impl(const pkmn::native::gen2_pc_pokemon_t& raw,
                              const pkmn::pkstring& nickname,
                              const pkmn::pkstring& otname,
-                             uint8_t version);
+                             int version);
             pokemon_gen2impl(const pkmn::native::gen2_party_pokemon_t& raw,
-                             uint8_t version);
+                             int version);
             pokemon_gen2impl(const pkmn::native::gen2_party_pokemon_t& raw,
                              const pkmn::pkstring& nickname,
                              const pkmn::pkstring& otname,
-                             uint8_t version);
+                             int version);
             pokemon_gen2impl(const pokemon_gen2impl& other);
             pokemon_gen2impl& operator=(const pokemon_gen2impl& other);
             ~pokemon_gen2impl() {};
@@ -60,7 +60,7 @@ namespace pkmn
             uint16_t get_trainer_secret_id() const;
             pkmn::pkstring get_ball() const;
             pkmn::pkstring get_original_game() const;
-            uint8_t get_met_level() const;
+            int get_met_level() const;
 
             // Setting Trainer Info
             void set_nickname(const pkmn::pkstring& nickname);
@@ -71,33 +71,33 @@ namespace pkmn
             void set_trainer_secret_id(uint16_t id);
             void set_ball(const pkmn::pkstring& ball);
             void set_original_game(const pkmn::pkstring& game);
-            void set_met_level(uint8_t level);
+            void set_met_level(int level);
 
             // Getting Individual Stat Info
             uint32_t get_personality() const;
-            uint8_t get_friendship() const;
-            uint8_t get_level() const;
+            int get_friendship() const;
+            int get_level() const;
             uint32_t get_experience() const;
             pkmn::pkstring get_gender() const;
             pkmn::nature_t get_nature() const;
             pkmn::pkstring get_ability() const;
             bool is_shiny() const;
-            pkmn::dict<pkmn::pkstring, uint16_t> get_stats() const;
-            pkmn::dict<pkmn::pkstring, uint16_t> get_EVs() const;
-            pkmn::dict<pkmn::pkstring, uint16_t> get_IVs() const;
+            pkmn::dict<pkmn::pkstring, int> get_stats() const;
+            pkmn::dict<pkmn::pkstring, int> get_EVs() const;
+            pkmn::dict<pkmn::pkstring, int> get_IVs() const;
             
             // Setting Individual Stat Info
             void set_personality(uint32_t personality);
-            void set_friendship(uint8_t friendship);
-            void set_level(uint8_t level);
+            void set_friendship(int friendship);
+            void set_level(int level);
             void set_experience(uint32_t experience);
             void set_gender(const pkmn::pkstring& gender);
             void set_nature(const pkmn::pkstring& nature_name);
             void set_ability(const pkmn::pkstring& ability);
             void set_form(const pkmn::pkstring& form);
             void set_shiny(bool value);
-            void set_EV(const pkmn::pkstring& stat, uint16_t value);
-            void set_IV(const pkmn::pkstring& stat, uint16_t value);
+            void set_EV(const pkmn::pkstring& stat, int value);
+            void set_IV(const pkmn::pkstring& stat, int value);
 
             // Battle Stat Info
             pkmn::pkstring get_status() const;
@@ -106,20 +106,20 @@ namespace pkmn
             void set_held_item(const pkmn::pkstring& item_name);
 
             // Getting Move Info
-            pkmn::move_entry_t get_move(uint8_t pos) const;
+            pkmn::move_entry_t get_move(int pos) const;
             void get_moves(pkmn::moveset_t& moves) const;
-            uint8_t get_move_PP(uint8_t pos) const;
-            void get_move_PPs(std::vector<uint8_t>& move_PPs) const;
+            int get_move_PP(int pos) const;
+            void get_move_PPs(std::vector<int>& move_PPs) const;
 
             // Setting Move Info
-            void set_move(const pkmn::pkstring& move_name, uint8_t pos);
-            void set_move_PP(uint8_t PP, uint8_t pos);
+            void set_move(const pkmn::pkstring& move_name, int pos);
+            void set_move_PP(int PP, int pos);
 
             // Database Info
-            uint16_t get_original_game_id() const;
-            uint16_t get_ability_id() const;
-            uint16_t get_item_id() const;
-            uint16_t get_nature_id() const;
+            int get_original_game_id() const;
+            int get_ability_id() const;
+            int get_item_id() const;
+            int get_nature_id() const;
 
             const void* get_native();
 
@@ -130,7 +130,7 @@ namespace pkmn
             pkmn::pkstring _nickname, _otname;
 
             void _set_experience(uint32_t exp);
-            void _set_level(uint8_t level);
+            void _set_level(int level);
             void _set_stats();
             void _reset_caught_data();
     };
